Replace PreCSS with sass[-loader]
PreCSS [1], which was introduced in [2], is to allow 'use[ing] Sass-like
markup and staged CSS features in CSS'.

However the original style sheet is already written in SCSS, so there is
no need to treat it as a CSS file and only use PreCSS to partially
support SCSS syntax.

Hence switching to scss & scss-loader. It also removes the ansi-regex
v2.1.1 dependency which had vulnerabilities [3].

[1] https://github.com/csstools/precss
[2] https://github.com/zendesk/v2_repl_app/commit/0f8a15bc2393fd02e45f72d5acbf34de5f5e9256
[3] https://app.snyk.io/vuln/SNYK-JS-ANSIREGEX-1583908